+++ packages/rust-bindings/src/external_functions.rs
@@ -55,6 +55,13 @@ pub fn create_id_request_v1_ext(input: &str) -> String {
     }
 }
 
+#[wasm_bindgen(js_name = createIdentityRecoveryRequest)]
+pub fn create_identity_recovery_request_ext(input: &str) -> String {
+    error_to_string(create_identity_recovery_request_aux(
+        serde_json::from_str(input).unwrap(),
+    ))
+}
+
 #[wasm_bindgen(js_name = createCredentialV1)]
 pub fn create_credential_v1_ext(raw_input: &str) -> String {
     match serde_json::from_str(raw_input) {