+++ packages/common/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `createIdentityRequest`, to create identity requests.
 - `createCredentialV1`, to create credentials using a seedPhrase.
+- `createIdentityRecoveryRequest`, to create identity recovery requests.
 - Added `sendCredentialDeployment` to send credentials created from `createCredentialV1` to the chain.
 - `getSignedCredentialDeploymentTransactionHash` to get the transaction hash of credentials created from `createCredentialV1`.
 - Added `ConfigureBaker` to `AccountTransactionType` enum.