+++ packages/rust-bindings/CHANGELOG.md
@@ -6,4 +6,5 @@
 
 - `createCredentialV1`
 - `createIdRequestV1`
+- `createIdentityRecoveryRequest`
 -  Bindings for the HdWallet methods: `getAccountSigningKey`, `getAccountPublicKey`, `getPrfKey`, `getSignatureBlindingRandomness` and `getAttributeCommitmentRandomness`.